Refactor packet and tracker classes for Monte Carlo simulation; implement RPacket and VPacket with associated methods
