+++ tardis/transport/montecarlo/packets/radiative_packet.py
